Exceptions Handled

+++ src/main/java/com/moorabi/reelsapi/controller/CommentController.java
@@ -10,9 +10,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.moorabi.reelsapi.exception.NotAllowedException;
 import com.moorabi.reelsapi.exception.ResourceNotFoundException;
 import com.moorabi.reelsapi.model.Comment;
 import com.moorabi.reelsapi.model.Reel;
@@ -20,6 +22,7 @@ import com.moorabi.reelsapi.model.User;
 import com.moorabi.reelsapi.repository.CommentRepository;
 import com.moorabi.reelsapi.repository.ReelRepository;
 import com.moorabi.reelsapi.repository.UserRepository;
+import com.moorabi.reelsapi.util.JwtTokenUtil;
 
 @RestController
 @RequestMapping("/api/v1")
@@ -33,37 +36,60 @@ public class CommentController {
 	@Autowired
 	private CommentRepository commentRepository;
 	
-	@GetMapping("/reels/{reel_id}")
+	@Autowired
+	private JwtTokenUtil jwtTokenUtil;
+	
+	@GetMapping("/reels/{reel_id}/comments")
 	public List<Comment> getAllCommentsForReel(@PathVariable(value="reel_id") long id){
-		Reel reel=reelRepository.getById(id);
+		Reel reel=reelRepository.findById(id).get();
 		return commentRepository.findAllForReel(reel);
 	}
 //	
-	@PostMapping("/reels/{reel_id}/users/{user_id}")
-	public Comment createComment(@PathVariable(value="user_id") long user_id,@PathVariable(value="reel_id") long reel_id,@RequestBody Comment comment) {
+	@PostMapping("/reels/{reel_id}/comment")
+	public Comment createComment(@RequestHeader (name="Authorization") String token,
+			@PathVariable(value="reel_id") long reel_id,@RequestBody Comment comment) {
+		String userName=jwtTokenUtil.getUsernameFromToken(token.split(" ")[1]);
 		Reel r=reelRepository.findById(reel_id).get();
-		r.getComments().add(comment);
 		comment.setReel(r);
-		comment.setUser(userRepository.findById(user_id).get());
-		reelRepository.save(r);
+		comment.setUser(userRepository.findUserByUsername(userName));
 		return commentRepository.save(comment);
 	}
 
 
-	@PutMapping("/reels/{reel_id}/users/{user_id}/comments/{comment_id}")
-	public ResponseEntity<Comment> updateComment(@PathVariable(value="user_id") long user_id,@PathVariable(value="reel_id") long reel_id,@PathVariable(value="comment_id") long comment_id,@RequestBody Comment commentDetails) throws ResourceNotFoundException {
+	@PutMapping("/reels/{reel_id}/comments/{comment_id}")
+	public ResponseEntity<Comment> updateComment(@RequestHeader (name="Authorization") String token,
+			@PathVariable(value="reel_id") long reel_id,
+			@PathVariable(value="comment_id") long comment_id,
+			@RequestBody Comment commentDetails) throws ResourceNotFoundException, NotAllowedException {
+		String userName=jwtTokenUtil.getUsernameFromToken(token.split(" ")[1]);
 		Comment c= commentRepository.findById(comment_id).orElseThrow(() -> new ResourceNotFoundException("Comment not found for this id : "+comment_id));
-//		Reel r=reelRepository.findById(reel_id).get();
+		User u=(userRepository.findById(c.getUser()))
+				.orElseThrow(() -> new ResourceNotFoundException("User not found for this user name : "+userName));;
+		if(!u.getUsername().equals(userName)) {
+			throw new NotAllowedException("Only Owner of reel can update it");
+		}
+		Reel r=reelRepository.findById(reel_id).get();
 		if(commentDetails.getDescription()!=null)
 			c.setDescription(commentDetails.getDescription());
 		commentRepository.save(c);
+		reelRepository.save(r);
 		return ResponseEntity.ok().body(c);
 	}
 	
-	@DeleteMapping("/reels/{reel_id}/users/{user_id}/comments/{comment_id}")
-	public ResponseEntity<?> deleteComment(@PathVariable(value="user_id") long user_id,@PathVariable(value="reel_id") long reel_id,@PathVariable(value="comment_id") long comment_id) throws ResourceNotFoundException {
-		Comment c= commentRepository.findById(comment_id).orElseThrow(() -> new ResourceNotFoundException("Comment not found for this id : "+comment_id));
+	@DeleteMapping("/reels/{reel_id}/comments/{comment_id}")
+	public ResponseEntity<?> deleteComment(@RequestHeader (name="Authorization") String token,
+			@PathVariable(value="reel_id") long reel_id,
+			@PathVariable(value="comment_id") long comment_id) throws ResourceNotFoundException, NotAllowedException {
+		String userName=jwtTokenUtil.getUsernameFromToken(token.split(" ")[1]);
+
+		Comment c= commentRepository.findById(comment_id)
+				.orElseThrow(() -> new ResourceNotFoundException("Comment not found for this id : "+comment_id));
 //		Reel r=reelRepository.findById(reel_id).get();
+		User u=(userRepository.findById(c.getUser()))
+				.orElseThrow(() -> new ResourceNotFoundException("User not found for this user name : "+userName));;
+		if(!u.getUsername().equals(userName)) {
+			throw new NotAllowedException("Only Owner of reel can delete it");
+		}
 		commentRepository.deleteById(comment_id);
 		return ResponseEntity.ok().build();
 	}
